PRSD-976: Give each class its own file
